Refactoring to move the model and interface packages out of the internal package
