+++ aws-serverless-java-container-core/src/main/java/com/amazonaws/serverless/proxy/internal/AwsProxyExceptionHandler.java
@@ -13,8 +13,10 @@
 package com.amazonaws.serverless.proxy.internal;
 
 import com.amazonaws.serverless.exceptions.InvalidRequestEventException;
-import com.amazonaws.serverless.proxy.internal.model.AwsProxyResponse;
-import com.amazonaws.serverless.proxy.internal.model.ErrorModel;
+import com.amazonaws.serverless.proxy.model.AwsProxyResponse;
+import com.amazonaws.serverless.proxy.model.ErrorModel;
+import com.amazonaws.serverless.proxy.ExceptionHandler;
+
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import org.slf4j.Logger;
@@ -34,7 +36,7 @@ import java.util.Map;
  * Returns application/json messages with a status code of 500 when the RequestReader failed to read the incoming event.
  * For all other exceptions returns a 502. Responses are populated with a JSON object containing a message property.
  *
- * @see com.amazonaws.serverless.proxy.internal.ExceptionHandler
+ * @see ExceptionHandler
  */
 public class AwsProxyExceptionHandler
         implements ExceptionHandler<AwsProxyResponse> {
